refactor(client): Remove all antenna mentions
