+++ clients/client/pom.xml
@@ -2,19 +2,22 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0"
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <parent>
-        <artifactId>clients</artifactId>
-        <groupId>org.eclipse.sw360</groupId>
-        <version>12.1.0-SNAPSHOT</version>
-    </parent>
     <modelVersion>4.0.0</modelVersion>
 
+    <parent>
+        <groupId>org.eclipse.sw360</groupId>
+        <artifactId>clients</artifactId>
+        <version>${revision}</version>
+    </parent>
+
     <artifactId>client</artifactId>
+    <packaging>jar</packaging>
 
     <dependencies>
         <dependency>
             <groupId>org.eclipse.sw360</groupId>
             <artifactId>http-support</artifactId>
+            <version>${project.version}</version>
         </dependency>
         <dependency>
             <groupId>org.slf4j</groupId>
@@ -39,6 +42,7 @@
         <dependency>
             <groupId>com.github.package-url</groupId>
             <artifactId>packageurl-java</artifactId>
+            <version>1.1.1</version>
         </dependency>
         <!-- Test Dependencies -->
         <dependency>
