+++ clients/client/src/main/java/org/eclipse/sw360/clients/auth/AccessToken.java
@@ -11,8 +11,8 @@
 package org.eclipse.sw360.clients.auth;
 
 import org.apache.commons.lang3.Validate;
-import org.eclipse.sw360.antenna.http.RequestBuilder;
-import org.eclipse.sw360.antenna.http.utils.HttpConstants;
+import org.eclipse.sw360.http.RequestBuilder;
+import org.eclipse.sw360.http.utils.HttpConstants;
 
 import java.util.Objects;
 import java.util.function.Consumer;
