+++ clients/client/src/main/java/org/eclipse/sw360/clients/utils/FutureUtils.java
@@ -10,9 +10,9 @@
  */
 package org.eclipse.sw360.clients.utils;
 
-import org.eclipse.sw360.antenna.http.utils.FailedRequestException;
-import org.eclipse.sw360.antenna.http.utils.HttpConstants;
-import org.eclipse.sw360.antenna.http.utils.HttpUtils;
+import org.eclipse.sw360.http.utils.FailedRequestException;
+import org.eclipse.sw360.http.utils.HttpConstants;
+import org.eclipse.sw360.http.utils.HttpUtils;
 
 import java.io.IOException;
 import java.util.Collection;
