+++ clients/client/src/test/java/org/eclipse/sw360/clients/adapter/SW360ComponentClientAdapterAsyncImplTest.java
@@ -10,8 +10,8 @@
  */
 package org.eclipse.sw360.clients.adapter;
 
-import org.eclipse.sw360.antenna.http.utils.FailedRequestException;
-import org.eclipse.sw360.antenna.http.utils.HttpConstants;
+import org.eclipse.sw360.http.utils.FailedRequestException;
+import org.eclipse.sw360.http.utils.HttpConstants;
 import org.eclipse.sw360.clients.rest.MultiStatusResponse;
 import org.eclipse.sw360.clients.rest.PagingResult;
 import org.eclipse.sw360.clients.rest.SW360ComponentClient;
