+++ clients/client/src/test/java/org/eclipse/sw360/clients/adapter/SW360ReleaseClientAdapterAsyncImplTest.java
@@ -13,8 +13,8 @@ package org.eclipse.sw360.clients.adapter;
 import com.github.packageurl.MalformedPackageURLException;
 import com.github.packageurl.PackageURL;
 import org.eclipse.sw360.clients.rest.SW360AttachmentAwareClient;
-import org.eclipse.sw360.antenna.http.utils.FailedRequestException;
-import org.eclipse.sw360.antenna.http.utils.HttpConstants;
+import org.eclipse.sw360.http.utils.FailedRequestException;
+import org.eclipse.sw360.http.utils.HttpConstants;
 import org.eclipse.sw360.clients.rest.MultiStatusResponse;
 import org.eclipse.sw360.clients.rest.SW360ReleaseClient;
 import org.eclipse.sw360.clients.rest.resource.LinkObjects;
