+++ clients/client/src/test/java/org/eclipse/sw360/clients/rest/SW360AttachmentClientIT.java
@@ -11,8 +11,8 @@
 package org.eclipse.sw360.clients.rest;
 
 import org.apache.commons.io.IOUtils;
-import org.eclipse.sw360.antenna.http.utils.FailedRequestException;
-import org.eclipse.sw360.antenna.http.utils.HttpConstants;
+import org.eclipse.sw360.http.utils.FailedRequestException;
+import org.eclipse.sw360.http.utils.HttpConstants;
 import org.eclipse.sw360.clients.rest.resource.Self;
 import org.eclipse.sw360.clients.rest.resource.attachments.SW360Attachment;
 import org.eclipse.sw360.clients.rest.resource.attachments.SW360AttachmentType;
@@ -46,7 +46,7 @@ import static com.github.tomakehurst.wiremock.client.WireMock.post;
 import static com.github.tomakehurst.wiremock.client.WireMock.postRequestedFor;
 import static com.github.tomakehurst.wiremock.client.WireMock.urlPathEqualTo;
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.eclipse.sw360.antenna.http.utils.HttpUtils.waitFor;
+import static org.eclipse.sw360.http.utils.HttpUtils.waitFor;
 
 public class SW360AttachmentClientIT extends AbstractMockServerTest {
