+++ clients/client/src/test/java/org/eclipse/sw360/clients/rest/SW360ComponentClientIT.java
@@ -11,8 +11,8 @@
 package org.eclipse.sw360.clients.rest;
 
 import org.assertj.core.api.Assertions;
-import org.eclipse.sw360.antenna.http.utils.FailedRequestException;
-import org.eclipse.sw360.antenna.http.utils.HttpConstants;
+import org.eclipse.sw360.http.utils.FailedRequestException;
+import org.eclipse.sw360.http.utils.HttpConstants;
 import org.eclipse.sw360.clients.rest.resource.Paging;
 import org.eclipse.sw360.clients.rest.resource.components.ComponentSearchParams;
 import org.eclipse.sw360.clients.rest.resource.components.SW360Component;
@@ -40,7 +40,7 @@ import static com.github.tomakehurst.wiremock.client.WireMock.patch;
 import static com.github.tomakehurst.wiremock.client.WireMock.post;
 import static com.github.tomakehurst.wiremock.client.WireMock.urlPathEqualTo;
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.eclipse.sw360.antenna.http.utils.HttpUtils.waitFor;
+import static org.eclipse.sw360.http.utils.HttpUtils.waitFor;
 
 public class SW360ComponentClientIT extends AbstractMockServerTest {
     /**
