+++ clients/client/src/test/java/org/eclipse/sw360/clients/rest/SW360ProjectClientIT.java
@@ -10,8 +10,8 @@
  */
 package org.eclipse.sw360.clients.rest;
 
-import org.eclipse.sw360.antenna.http.utils.FailedRequestException;
-import org.eclipse.sw360.antenna.http.utils.HttpConstants;
+import org.eclipse.sw360.http.utils.FailedRequestException;
+import org.eclipse.sw360.http.utils.HttpConstants;
 import org.eclipse.sw360.clients.rest.resource.projects.ProjectSearchParams;
 import org.eclipse.sw360.clients.rest.resource.projects.SW360Project;
 import org.eclipse.sw360.clients.rest.resource.projects.SW360ProjectType;
@@ -37,7 +37,7 @@ import static com.github.tomakehurst.wiremock.client.WireMock.post;
 import static com.github.tomakehurst.wiremock.client.WireMock.postRequestedFor;
 import static com.github.tomakehurst.wiremock.client.WireMock.urlPathEqualTo;
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.eclipse.sw360.antenna.http.utils.HttpUtils.waitFor;
+import static org.eclipse.sw360.http.utils.HttpUtils.waitFor;
 
 public class SW360ProjectClientIT extends AbstractMockServerTest {
     /**
