+++ clients/http-support/pom.xml
@@ -2,14 +2,16 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0"
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <parent>
-        <artifactId>clients</artifactId>
-        <groupId>org.eclipse.sw360</groupId>
-        <version>12.1.0-SNAPSHOT</version>
-    </parent>
     <modelVersion>4.0.0</modelVersion>
 
+    <parent>
+        <groupId>org.eclipse.sw360</groupId>
+        <artifactId>clients</artifactId>
+        <version>${revision}</version>
+    </parent>
+
     <artifactId>http-support</artifactId>
+    <packaging>jar</packaging>
 
     <dependencies>
         <dependency>
