+++ clients/pom.xml
@@ -8,20 +8,22 @@
   ~
   ~ SPDX-License-Identifier: EPL-2.0
   -->
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+
     <parent>
-        <artifactId>sw360</artifactId>
         <groupId>org.eclipse.sw360</groupId>
+        <artifactId>sw360</artifactId>
         <version>${revision}</version>
     </parent>
 
     <artifactId>clients</artifactId>
     <packaging>pom</packaging>
+
     <modules>
-        <module>client</module>
         <module>http-support</module>
+        <module>client</module>
     </modules>
 
 </project>
