+++ pom.xml
@@ -120,6 +120,10 @@
                 <artifactId>json-simple</artifactId>
                 <version>2.3.1</version>
             </dependency>
+            <dependency>
+                <groupId>com.github.package-url</groupId>
+                <artifactId>packageurl-java</artifactId>
+            </dependency>
             <!-- javax.annotation: java 11 compatibility -->
             <dependency>
               <groupId>javax.annotation</groupId>
@@ -413,6 +417,12 @@
                 <groupId>com.github.tomakehurst</groupId>
                 <artifactId>wiremock-jre8</artifactId>
                 <version>${wiremock.version}</version>
+                <scope>test</scope>
+            </dependency>
+            <dependency>
+                <groupId>org.slf4j</groupId>
+                <artifactId>slf4j-simple</artifactId>
+                <version>${slf4j.version}</version>
                 <scope>test</scope>
             </dependency>
         </dependencies>
